Update puppet_master_2501.py
Added logging capabilities.

+++ puppet_master_2501.py
@@ -1,123 +1,195 @@
 #!/usr/bin/python3
 
 import argparse
+import logging
 import os
 import shutil
+
+
+print("Setting up logging...")
+log_path = os.path.join(os.getcwd(), "logs/puppet_master.log")
+if not os.path.isdir("logs"):
+    print("Logging directory not found.")
+    print("Creating logging directory...")
+    log_dir_path = os.path.join(os.getcwd(), "logs")
+    os.mkdir(log_dir_path)
+    f = open(log_path, "x")
+    f.close()
+
+elif not os.path.exists(log_path):
+    f = open(log_path, "x")
+    f.close()
+
+logging.basicConfig(
+    level=logging.INFO,
+    format="%(asctime)s - %(name)s - %(levelname)s - %(message)s",
+    filename=log_path)
+
+logger = logging.getLogger()
 
 
 def copy_file(filename, dirname):
     """
     """
 
+    print("Copying file...")
+    logger.info("Copying file...")
     src = filename
     dest = dirname
-    shutil.copy(src, dest)
-    print("{0} has been copied.".format(filename))
+    if os.path.isfile(filename):
+        shutil.copy(src, dest)
+        print("{0} has been copied.".format(filename))
+        logger.info("{0} has been copied.".format(filename))
+
+    else:
+        print("What do you mean you're looking at him right now?")
+        logger.info("What do you mean you're looking at him right now?")
+        print("{0} is already there.".format(filename))
+        logger.info("{0} is already there.".format(filename))
 
 
 def create_dir(dirname):
     """
     """
 
+    print("Creating directory...")
+    logger.info("Creating directory...")
     if not os.path.isdir(dirname):
         path = os.path.join(os.getcwd(), dirname)
         os.mkdir(path)
         print("{0} has been created.".format(dirname))
+        logger.info("{0} has been created.".format(dirname))
 
     else:
         print("Bro, what is you doin'?")
+        logger.info("I think we already got him, Jet.")
         print("{0} already exists.".format(dirname))
+        logger.info("{0} already exists.".format(dirname))
 
 
 def create_file(filename, content):
     """
     """
 
+    print("Creating file...")
+    logger.info("Creating file...")
     if not os.path.exists(filename):
         f = open(filename, "w")
         f.write(content)
         f.close()
         print("{0} has been created.".format(filename))
+        logger.info("{0} has been created.".format(filename))
 
     else:
         print("Tell me something I don't know.")
+        logger.info("Tell me something I don't know.")
         print("{0} already exists.".format(filename))
+        logger.info("{0} already exists.".format(filename))
 
 
 def delete_dir(dirname):
     """
     """
 
+    print("Deleting directory...")
+    logger.info("Deleting directory...")
     if os.path.isdir(dirname):
         os.rmdir(dirname)
         print("{0} has been deleted.".format(dirname))
+        logger.info("{0} has been deleted.".format(dirname))
 
     else:
         print("{0} does not exist.".format(dirname))
+        logger.info("{0} does not exist.".format(dirname))
 
 
 def delete_file(filename):
     """
     """
 
+    print("Deleting file...")
+    logger.info("Deleting file...")
     if os.path.isfile(filename):
         os.remove(filename)
         print("{0} has been deleted.".format(filename))
+        logger.info("{0} has been deleted.".format(filename))
 
     else:
         print("Who?")
+        logger.info("Who?")
         print("{0} does not exists.".format(filename))
+        logger.info("{0} does not exists.".format(filename))
 
 
 def find_file(filename, search_path):
 
+    print("Looking for file...")
+    logger.info("Looking for file...")
     result = []
     for root, dir, files in os.walk(search_path):
         if filename in files:
             print("Found it!")
+            logger.info("Found it!")
             print(result.append(os.path.join(root, filename)))
+            logger.info(result.append(os.path.join(root, filename)))
 
         elif filename not in files:
             print("Call off the search.")
+            logger.info("Call off the search.")
             print("{0} could not be found.".format(filename))
+            logger.info("{0} could not be found.".format(filename))
 
 
 def rename_dir(olddirname, newdirname):
     """
     """
 
+    print("Renaming directory...")
+    logger.info("Renaming directory...")
     source = olddirname
     dest = newdirname
     if os.path.exists(olddirname):
         os.rename(source, dest)
         print("This directory has been renamed.")
+        logger.info("This directory has been renamed.")
 
     elif not os.path.exists(olddirname):
         print("Let me see those directions.")
+        logger.info("Let me see those directions.")
         print("{0} cannot be found.".format(olddirname))
+        logger.info("{0} cannot be found.".format(olddirname))
 
     elif os.path.exists(newdirname):
         print("Let me see those directions.")
+        logger.info("Let me see those directions.")
         print("{0} is somewhere else.".format(newdirname))
+        logger.info("{0} is somewhere else.".format(newdirname))
 
 
 def rename_file(oldfilename, newfilename):
     """
     """
 
+    print("Renaming file...")
+    logger.info("Renaming file...")
     source = oldfilename
     dest = newfilename
     if os.path.exists(oldfilename):
         os.rename(source, dest)
         print("This file has been renamed.")
+        logger.info("This file has been renamed.")
 
     elif not os.path.exists(oldfilename):
         print("Let me see those directions.")
+        logger.info("Let me see those directions.")
         print("{0} cannot be found.".format(oldfilename))
+        logger.info("{0} cannot be found.".format(oldfilename))
 
     elif os.path.exists(newfilename):
         print("Let me see those directions.")
+        logger.info("Let me see those directions.")
         print("{0} is somewhere else.")
+        logger.info("{0} is somewhere else.")
 
 
 def main():
